A few changes to the attempts API.
Attempt runs now have a AttemptResult type and I removed the
function that takes a vector of attempts. The idea was flawed as
I would need Boxed values for the vector to use dynamic dispatch.
So the API is now only able to run a single attempt many times in
a row and then returns the average AttemptResult. However if the
found pairs are inconsistent between runs it will return Err(&str).